Test for Trie.Find
- rename test_trie -> trie_test
